Openshift: remove dnsmasq
DNSmasq process was not handled by Apicast at all, dnsmasq process can
die and
will never be up, so dns queries will start to fail.

The reasons to have APICast were:

- Openshift cluster provided multiple DNS servers in the resolv.conf:
not anymore, currently is using a Kubernetes service with static IP, so
all connections goes to the same service.
- DNS service was block during pod updates, no longer a case, it's
working correctly, added more pods in to a service and do not have the
case.
- Cache was not used, no longer needed due to `resolver.search_dns` is
looking for the full query in the cache before make the dns request to
the server. [0]

[0]
https://github.com/3scale/APIcast/blob/41ad2c2a04054cd05991a65fc807a8105cdb2fd6/gateway/src/resty/resolver.lua#L282-L302

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>