Openresty: Use openresty packages from approved build system.
This updates the version of Openresty to the version approved by the
build system.

Also create a new package for Opentracing that integrates
all in the same place.

Also, using the new features added in the master branch to support
Openssl 1.1.1 natively to be FIPS compliance.

Move the container to Centos 8 version.

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>